+++ packages/cubejs-client-core/src/RequestError.js
@@ -1,0 +1,6 @@
+export default class RequestError extends Error {
+  constructor(message, response) {
+    super(message);
+    this.response = response;
+  }
+}